perf(frontend): load the editor and emoji picker on demand
The build page statically imported a 1335 kB chunk (306 kB gzip) holding
TipTap/ProseMirror, the emojibase dataset and motion. All of it was fetched and
parsed before the page could paint, ahead of the screenshot diffs that are the
reason to open a build at all — and whether or not the reader ever looked at a
comment, let alone wrote one.

Split both behind lazy boundaries, keeping their public APIs unchanged so no
call site moves:

- `Editor` becomes a façade over `EditorImpl`, so the ~450 kB of TipTap and
  ProseMirror load when a comment surface first renders. `ReadOnlyEditor` shares
  the same editor, so this defers displaying comments as well as composing them.
- `EmojiPicker` becomes a façade over `EmojiPickerGrid`, moving 571 kB of
  emojibase JSON — and the index it builds at module scope — behind the click
  that opens the picker.

Both fallbacks reserve the real footprint and carry `aria-busy`, so nothing
reflows and Argos waits for the resolved state instead of screenshotting a gap.

Build page static graph: 1335 kB -> 317 kB raw, 306 kB -> 87 kB gzip. The
deferred code splits into EditorImpl (450 kB / 134 kB gzip) and EmojiPickerGrid
(570 kB / 86 kB gzip), neither statically imported by any chunk.
